add museums distance view

+++ database/migrations/2023_06_14_142839_create_museums_table.php
@@ -11,27 +11,27 @@ return new class extends Migration
      *
      * @return void
      */
-    public function up()
-    {
-        Schema::create('museums', function (Blueprint $table) {
-            $table->id();
+  public function up()
+  {
+    Schema::create('museums', function (Blueprint $table) {
+      $table->id();
 
-            $table->string('name', 50);
-            $table->integer('coordinate_x');
-            $table->integer('coordinate_y');
-            $table->string('city', 100);
+      $table->string('name', 50);
+      $table->decimal('coordinate_x', 11, 8);
+      $table->decimal('coordinate_y', 11, 8);
+      $table->string('city', 100);
 
-            $table->timestamps();
-        });
-    }
+      $table->timestamps();
+    });
+  }
 
-    /**
-     * Reverse the migrations.
-     *
-     * @return void
-     */
-    public function down()
-    {
-        Schema::dropIfExists('museums');
-    }
+  /**
+   * Reverse the migrations.
+   *
+   * @return void
+   */
+  public function down()
+  {
+      Schema::dropIfExists('museums');
+  }
 };
